+++ package.json
@@ -10,7 +10,7 @@
     "test:unit": "cross-env tape 'test/**/*-test.js' | tap-spec",
     "coverage": "nyc --reporter=lcov --reporter=text-summary npm run test:unit",
     "test": "npm run lint && npm run coverage",
-    "t": "npm run lint && tape test/02-parser-test.js | tap-spec",
+    "t": "npm run lint && tape test/12-schema-test.js | tap-spec",
     "rc": "npm version prerelease --preid RC",
     "build": "rollup -c",
     "vendor:js-yaml": "mkdir -p tmp && browserify --node --standalone js-yaml --im --no-builtins node_modules/js-yaml/index.js > tmp/js-yaml.js && terser tmp/js-yaml.js -o src/compat/vendor/js-yaml.min.js",
