+++ test/06-compat-test.js
