--- test/06-scheduled-test.js
@@ -1,11 +1,0 @@
-var test = require('tape')
-var fs = require('fs')
-var parse = require('../')
-
-test('@scheduled', t => {
-  t.plan(1)
-  var mock = fs.readFileSync('./test/mock/scheduled.arc').toString()
-  var parsed = parse(mock)
-  t.ok(parsed, 'read out scheduled')
-  console.log(parsed)
-})
