--- test/mock/invalid.arc
@@ -1,2 +1,0 @@
-#nope
-asdfasdfa
