--- test/mock/invalid2.arc
@@ -1,2 +1,0 @@
-@sok sun
-asdfasdfa
