--- test/mock/invalid3.arc
@@ -1,3 +1,0 @@
-@sok
-nice nice
-  he he he
