--- test/mock/invalid4.arc
@@ -1,3 +1,0 @@
-@wut
-  he he he
-  he he he
